assembler, runner overhaul with tons of bashism
From now on, all env/vars cannot contain any shell expansion, i.e., they
are used as static strings to define the environment or shell variables.

Not sure if it's good to keep the two worlds of env/vars and force
bashism to run.sh as well.  Maybe it's better to only use environment
variables and replace run.sh to be an arbitrary executable, i.e.,
sticking to the basic Unix-style.