Relax the strict file permissions made by mktemp
mktemp always creates temporary files/dirs with go= permission.  When
sharing the repo, this can be problematic.  As a workaround, using the
same permission as the repo seems okay for now.  For completeness, we
may need to copy the repo root's ACL, and provide more clear ways to
share the repo among a group of users.